requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
cloudpickle==3.1.1
colorama==0.4.6
deap==1.4.2
Django==5.1.6
filelock==3.17.0
flatbuffers==25.2.10
fsspec==2025.2.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
imbalanced-learn==0.13.0
Jinja2==3.1.5
joblib==1.4.2
keras==3.9.0
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.2
shap==0.47.0
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
sqlparse==0.5.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
threadpoolctl==3.5.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
cloudpickle==3.1.1
colorama==0.4.6
deap==1.4.2
Django==5.1.6
filelock==3.17.0
flatbuffers==25.2.10
fsspec==2025.2.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
imbalanced-learn==0.13.0
Jinja2==3.1.5
joblib==1.4.2
keras==3.9.0
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.2
shap==0.47.0
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
sqlparse==0.5.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
threadpoolctl==3.5.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
